refactor(marketing): move the cadence handler out of packs/ (BI-C26FE785)
The Tool Surface Guard failed: registry.packFiles grew 98 -> 99. Every file in
lib/mcp/packs/ counts as a pack, whether or not it defines tools, and the
extracted handler defines none — registry.toolDefinitions was unchanged at 405.

Re-baselining was the guard's suggested path, but its prompt asks the author to
"record WHY this tool earns a permanent place in the selection space". No tool
was added, so answering that honestly was impossible; raising the baseline
would have recorded a claim that is not true and left the pack count a slightly
worse proxy for the surface an agent must discriminate within.

Moved to lib/mcp/marketing-cadence-handler.ts instead, beside the eight
existing lib/mcp/*-handler(s).ts modules (build-design-review-handler,
deliberation-handlers, build-lifecycle-handlers and the rest). That is the
established home for MCP handler logic that is not itself a pack. Tool surface
is back to 405 definitions across 98 packs with no baseline change.

No behaviour change: 44 test files / 353 tests pass.

Design-Grounding-Decision: No contract change. File relocation only, to keep the tool-surface metric measuring what it claims to measure. Substrate reviewed: scripts/check-tool-surface.mjs measureToolSurface counts one pack per file in lib/mcp/packs regardless of tool count; the lib/mcp/*-handler(s).ts convention already exists for exactly this case.
